% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_infoUI_zocalo_03_container.R
\name{fn_infoUI_zocalo_03_container}
\alias{fn_infoUI_zocalo_03_container}
\title{Función auxiliar para generar el contenedor UI del zócalo de metadatos.}
\usage{
fn_infoUI_zocalo_03_container(dt_output_id)
}
\arguments{
\item{dt_output_id}{El ID del objeto DTOutput (ej. "settings_table_display")
que se incrustará dentro del zócalo.}
}
\value{
Un objeto de tagList (código HTML/Shiny UI).
}
\description{
Función auxiliar para generar el contenedor UI del zócalo de metadatos.
}
