% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_reactive_values_from_list.R
\name{set_reactive_values_from_list}
\alias{set_reactive_values_from_list}
\title{set_reactive_values_from_list}
\usage{
set_reactive_values_from_list(rv, data_list)
}
\arguments{
\item{rv}{Un objeto reactiveValues de Shiny (pasado por referencia).}

\item{data_list}{Una lista R estándar cuyos elementos se usarán para actualizar rv.}
}
\value{
El objeto rv modificado (modifica por referencia, pero se devuelve para consistencia).
}
\description{
Sobrescribe los valores de un objeto reactiveValues (rv)
con los valores proporcionados en una lista (data_list).
Solo actualiza las claves que existen en la lista de datos.
}
