% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_infoUI_zocalo_03_container.R
\name{fn_infoUI_zocalo_03_container}
\alias{fn_infoUI_zocalo_03_container}
\title{Función auxiliar para generar el contenedor UI del zócalo de metadatos.}
\usage{
fn_infoUI_zocalo_03_container(data_obj, width = "100\%", height = "100\%")
}
\arguments{
\item{data_obj}{Objeto de datos con la información a mostrar}

\item{width}{Ancho del contenedor (ej: "100\%", "500px", "50vw")}

\item{height}{Alto del contenedor (ej: "100\%", "300px", "50vh")}
}
\value{
Un objeto de tagList (código HTML/Shiny UI).
}
\description{
Función auxiliar para generar el contenedor UI del zócalo de metadatos.
}
